// packages/openapi-generator/dist/typescript-node/model/allOfTest.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Message } from './message';

export class AllOfTest {
    /**
    * Number of items returned for each page.
    */
    'pageSize'?: number;
    /**
    * Contains an URL to GET the next #<page_size> results in the retrieved collection of items.
    */
    'next'?: string;
    'items'?: Array<Message>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "pageSize",
            "baseName": "page_size",
            "type": "number"
        },
        {
            "name": "next",
            "baseName": "next",
            "type": "string"
        },
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<Message>"
        }    ];

    static getAttributeTypeMap() {
        return AllOfTest.attributeTypeMap;
    }
}