// packages/openapi-generator/dist/typescript-node/model/allOfWithXExtensibleEnum.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* test if allOf with x-extensible-enum works fine
*/
export class AllOfWithXExtensibleEnum {
    'email'?: string;
    'familyName': string;
    /**
    * User\'s fiscal code.
    */
    'fiscalCode': string;
    'hasProfile': boolean;
    'isEmailSet': boolean;
    /**
    * True if the recipient of a message wants to store its content for later retrieval.
    */
    'isInboxEnabled'?: boolean = false;
    /**
    * True if the recipient of a message wants to forward the notifications to the default webhook.
    */
    'isWebhookEnabled'?: boolean = false;
    'name': string;
    'preferredEmail'?: string;
    /**
    * Indicates the User\'s preferred written or spoken languages in order of preference. Generally used for selecting a localized User interface. Valid values are concatenation of the ISO 639-1 two letter language code, an underscore, and the ISO 3166-1 2 letter country code; e.g., \'en_US\' specifies the language English and country US.
    */
    'preferredLanguages'?: Array<string>;
    'version': number;
    'payload'?: object;
    'status': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "familyName",
            "baseName": "family_name",
            "type": "string"
        },
        {
            "name": "fiscalCode",
            "baseName": "fiscal_code",
            "type": "string"
        },
        {
            "name": "hasProfile",
            "baseName": "has_profile",
            "type": "boolean"
        },
        {
            "name": "isEmailSet",
            "baseName": "is_email_set",
            "type": "boolean"
        },
        {
            "name": "isInboxEnabled",
            "baseName": "is_inbox_enabled",
            "type": "boolean"
        },
        {
            "name": "isWebhookEnabled",
            "baseName": "is_webhook_enabled",
            "type": "boolean"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "preferredEmail",
            "baseName": "preferred_email",
            "type": "string"
        },
        {
            "name": "preferredLanguages",
            "baseName": "preferred_languages",
            "type": "Array<string>"
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "number"
        },
        {
            "name": "payload",
            "baseName": "payload",
            "type": "object"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AllOfWithXExtensibleEnum.attributeTypeMap;
    }
}