// packages/openapi-generator/dist/typescript-node/model/anObjectWithAnItemsField.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DefinitionFieldWithDash } from './definitionFieldWithDash';

/**
* What if a object has a field named items? The case is an object like { items: [] }, which is legal
*/
export class AnObjectWithAnItemsField {
    'items': Array<DefinitionFieldWithDash>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "items",
            "baseName": "items",
            "type": "Array<DefinitionFieldWithDash>"
        }    ];

    static getAttributeTypeMap() {
        return AnObjectWithAnItemsField.attributeTypeMap;
    }
}