// packages/openapi-generator/dist/typescript-node/model/anObjectWithRefImport.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { SimpleDefinition } from './simpleDefinition';

/**
* Describes an object with a ref import
*/
export class AnObjectWithRefImport {
    'prop1': SimpleDefinition;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "prop1",
            "baseName": "prop1",
            "type": "SimpleDefinition"
        }    ];

    static getAttributeTypeMap() {
        return AnObjectWithRefImport.attributeTypeMap;
    }
}