// packages/openapi-generator/dist/typescript-node/model/book.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { BookAuthor } from './bookAuthor';

export class Book {
    /**
    * title of the book
    */
    'title'?: string;
    'author'?: BookAuthor;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "author",
            "baseName": "author",
            "type": "BookAuthor"
        }    ];

    static getAttributeTypeMap() {
        return Book.attributeTypeMap;
    }
}