// packages/openapi-generator/dist/typescript-node/model/definitionFieldWithDash.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DefinitionFieldWithDash {
    'idField'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "idField",
            "baseName": "id-field",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DefinitionFieldWithDash.attributeTypeMap;
    }
}