// packages/openapi-generator/dist/typescript-node/model/disabledUserTest.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DisabledUserTest {
    'enabled': DisabledUserTest.EnabledEnum;
    'reason': string;
    'username': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "enabled",
            "baseName": "enabled",
            "type": "DisabledUserTest.EnabledEnum"
        },
        {
            "name": "reason",
            "baseName": "reason",
            "type": "string"
        },
        {
            "name": "username",
            "baseName": "username",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DisabledUserTest.attributeTypeMap;
    }
}

export namespace DisabledUserTest {
    export enum EnabledEnum {
        False = <any> false
    }
}