// packages/openapi-generator/dist/typescript-node/model/disjointUnionsUserTest.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DisabledUserTest } from './disabledUserTest';
import { EnabledUserTest } from './enabledUserTest';

export class DisjointUnionsUserTest {
    'description': string;
    'enabled': DisjointUnionsUserTest.EnabledEnum;
    'username': string;
    'reason': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "enabled",
            "baseName": "enabled",
            "type": "DisjointUnionsUserTest.EnabledEnum"
        },
        {
            "name": "username",
            "baseName": "username",
            "type": "string"
        },
        {
            "name": "reason",
            "baseName": "reason",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DisjointUnionsUserTest.attributeTypeMap;
    }
}

export namespace DisjointUnionsUserTest {
    export enum EnabledEnum {
        False = <any> false
    }
}