// packages/openapi-generator/dist/typescript-node/model/enabledUserTest.ts
/**
 * Test API
 * Test API.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class EnabledUserTest {
    'description': string;
    'enabled': EnabledUserTest.EnabledEnum;
    'username': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "enabled",
            "baseName": "enabled",
            "type": "EnabledUserTest.EnabledEnum"
        },
        {
            "name": "username",
            "baseName": "username",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return EnabledUserTest.attributeTypeMap;
    }
}

export namespace EnabledUserTest {
    export enum EnabledEnum {
        True = <any> true
    }
}